# cmd/api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.AdminListTokensResponse:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/api.AdminToken'
        type: array
      limit:
        type: integer
      next:
        type: integer
    type: object
...
  api.AdminToken:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      hash_algorithm:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      renewed_at:
        type: string
      revoked_at:
        type: string
      rotated_at:
        type: string
      type:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: User-facing read API for Prism candidate/content data.
  title: Prism API
  version: "0.1"
paths:
...
  /admin/tokens:
    get:
      parameters:
      - description: Page size (default 50, max 500)
        in: query
        name: limit
        type: integer
      - description: Cursor for next page (default 1)
        in: query
        name: next
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdminListTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List admin tokens
      tags:
      - admin
  /admin/tokens/{id}:
    get:
      parameters:
      - description: Token UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdminToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get admin token
      tags:
      - admin
...
swagger: "2.0"